Split ajv-error and error handling
